backend/requirements.txt
asgiref==3.8.1
astroid==3.3.9
certifi==2025.1.31
charset-normalizer==3.4.1
dill==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
networkx==3.4.2
packaging==24.2
platformdirs==4.3.6
psycopg2-binary==2.9.10
PyJWT==2.9.0
pylint==3.3.5
pylint-django==2.6.1
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
tomlkit==0.13.2
urllib3==2.3.0
